+++ src/main/java/pl/thedeem/intellij/common/sdk/errors/DQLApiException.java
@@ -1,9 +1,0 @@
-package pl.thedeem.intellij.common.sdk.errors;
-
-import java.io.IOException;
-
-public class DQLApiException extends IOException {
-    public DQLApiException(String message) {
-        super(message);
-    }
-}