+++ src/main/java/pl/thedeem/intellij/common/sdk/errors/DTAuthException.java
@@ -1,0 +1,11 @@
+package pl.thedeem.intellij.common.sdk.errors;
+
+public class DTAuthException extends DTIoException {
+    public DTAuthException(String message) {
+        super(message);
+    }
+
+    public DTAuthException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}