+++ src/main/java/pl/thedeem/intellij/common/sdk/errors/SSONotConfiguredException.java
@@ -1,0 +1,7 @@
+package pl.thedeem.intellij.common.sdk.errors;
+
+public class SSONotConfiguredException extends SSOAuthException {
+    public SSONotConfiguredException(String message) {
+        super(message);
+    }
+}