+++ src/main/java/pl/thedeem/intellij/common/sdk/errors/SSOReAuthRequiredException.java
@@ -1,0 +1,7 @@
+package pl.thedeem.intellij.common.sdk.errors;
+
+public class SSOReAuthRequiredException extends SSOAuthException {
+    public SSOReAuthRequiredException(String message) {
+        super(message);
+    }
+}